fix(ci): python-coverage-shards — split worker cleanup.py under 1500-line guardrail
cleanup.py grew to 1558 lines with the #582/#590 service-GC-trigger work,
tripping test_first_party_code_files_stay_under_line_limit (max 1500). Extract
the cohesive on-demand service_gc_requests trigger cluster
(_maybe_consume_service_gc_trigger, _claim_service_gc_trigger,
_reclaim_stale_running_service_gc_triggers, _run_claimed_service_gc_trigger,
_finish_service_gc_trigger) into a sibling cleanup_service_gc.py, mirroring the
existing cleanup_auth_overlay.py split, and re-wire the delegating mixin. Pure
code-move, behavior unchanged; cleanup.py now 1331 lines.
