+++ src/awf/control/worker/mixins.py
@@ -5,6 +5,7 @@ from __future__ import annotations
 from awf.control.worker import claims as _claims
 from awf.control.worker import cleanup as _cleanup
 from awf.control.worker import cleanup_auth_overlay as _cleanup_auth_overlay
+from awf.control.worker import cleanup_service_gc as _cleanup_service_gc
 from awf.control.worker import dispatch_methods as _dispatch_methods
 from awf.control.worker import recovery as _recovery
 from awf.control.worker import scheduler_methods as _scheduler_methods
@@ -44,11 +45,13 @@ class WorkerDelegatesMixin:
     _maybe_reap_classified_orphans = _cleanup._maybe_reap_classified_orphans
     _maybe_reap_superseded_claude_bases = _cleanup._maybe_reap_superseded_claude_bases
     _maybe_reap_terminal_workspace_gc = _cleanup._maybe_reap_terminal_workspace_gc
-    _maybe_consume_service_gc_trigger = _cleanup._maybe_consume_service_gc_trigger
-    _reclaim_stale_running_service_gc_triggers = _cleanup._reclaim_stale_running_service_gc_triggers
-    _claim_service_gc_trigger = _cleanup._claim_service_gc_trigger
-    _run_claimed_service_gc_trigger = _cleanup._run_claimed_service_gc_trigger
-    _finish_service_gc_trigger = _cleanup._finish_service_gc_trigger
+    _maybe_consume_service_gc_trigger = _cleanup_service_gc._maybe_consume_service_gc_trigger
+    _reclaim_stale_running_service_gc_triggers = (
+        _cleanup_service_gc._reclaim_stale_running_service_gc_triggers
+    )
+    _claim_service_gc_trigger = _cleanup_service_gc._claim_service_gc_trigger
+    _run_claimed_service_gc_trigger = _cleanup_service_gc._run_claimed_service_gc_trigger
+    _finish_service_gc_trigger = _cleanup_service_gc._finish_service_gc_trigger
     _release_terminal_runtime_resources = _cleanup._release_terminal_runtime_resources
     _resume_pending_planning_scope_auto_retries_after_terminal_release = (
         _cleanup._resume_pending_planning_scope_auto_retries_after_terminal_release
